backend/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
python-dotenv
scikit-learn
joblib
langchain
langchain-groq
httpx
